+++ src/LiveChartsCore/LineSeries.cs
@@ -256,11 +256,11 @@ namespace LiveChartsCore
                         v.Geometry.Width = gs;
                         v.Geometry.Height = gs;
 
-                        v.Bezier.X0 = x0b;
+                        v.Bezier.X0 = (float)x0b;
                         v.Bezier.Y0 = y0b;
-                        v.Bezier.X1 = x1b;
+                        v.Bezier.X1 = (float)x1b;
                         v.Bezier.Y1 = y1b;
-                        v.Bezier.X2 = x2b;
+                        v.Bezier.X2 = (float)x2b;
                         v.Bezier.Y2 = y2b;
 
                         data.TargetPoint.Context.Visual = v;
@@ -274,12 +274,12 @@ namespace LiveChartsCore
                     if (GeometryFill is not null) GeometryFill.AddGeometryToPaintTask(cartesianChart.Canvas, visual.Geometry);
                     if (GeometryStroke is not null) GeometryStroke.AddGeometryToPaintTask(cartesianChart.Canvas, visual.Geometry);
 
-                    visual.Bezier.X0 = data.X0;
-                    visual.Bezier.Y0 = data.Y0;
-                    visual.Bezier.X1 = data.X1;
-                    visual.Bezier.Y1 = data.Y1;
-                    visual.Bezier.X2 = data.X2;
-                    visual.Bezier.Y2 = data.Y2;
+                    visual.Bezier.X0 = (float)data.X0;
+                    visual.Bezier.Y0 = (float)data.Y0;
+                    visual.Bezier.X1 = (float)data.X1;
+                    visual.Bezier.Y1 = (float)data.Y1;
+                    visual.Bezier.X2 = (float)data.X2;
+                    visual.Bezier.Y2 = (float)data.Y2;
 
                     if (Fill is not null)
                     {
@@ -287,10 +287,10 @@ namespace LiveChartsCore
                         {
                             if (wasFillInitialized)
                             {
-                                fillPathHelper.StartPoint.X = data.X0;
+                                fillPathHelper.StartPoint.X = (float)data.X0;
                                 fillPathHelper.StartPoint.Y = p;
 
-                                fillPathHelper.StartSegment.X = data.X0;
+                                fillPathHelper.StartSegment.X = (float)data.X0;
                                 fillPathHelper.StartSegment.Y = p;
 
                                 fillPathHelper.StartPoint.CompleteTransitions(
@@ -299,12 +299,12 @@ namespace LiveChartsCore
                                     nameof(fillPathHelper.StartSegment.Y), nameof(fillPathHelper.StartSegment.X));
                             }
 
-                            fillPathHelper.StartPoint.X = data.X0;
+                            fillPathHelper.StartPoint.X = (float)data.X0;
                             fillPathHelper.StartPoint.Y = p;
                             fillPathHelper.Path.AddCommand(fillPathHelper.StartPoint);
 
-                            fillPathHelper.StartSegment.X = data.X0;
-                            fillPathHelper.StartSegment.Y = data.Y0;
+                            fillPathHelper.StartSegment.X = (float)data.X0;
+                            fillPathHelper.StartSegment.Y = (float)data.Y0;
                             fillPathHelper.Path.AddCommand(fillPathHelper.StartSegment);
                         }
 
@@ -312,7 +312,7 @@ namespace LiveChartsCore
 
                         if (data.IsLast)
                         {
-                            fillPathHelper.EndSegment.X = data.X2;
+                            fillPathHelper.EndSegment.X = (float)data.X2;
                             fillPathHelper.EndSegment.Y = p;
                             fillPathHelper.Path.AddCommand(fillPathHelper.EndSegment);
 
@@ -334,7 +334,7 @@ namespace LiveChartsCore
                                 }
                                 else
                                 {
-                                    strokePathHelper.StartPoint.X = data.X0;
+                                    strokePathHelper.StartPoint.X = (float)data.X0;
                                     strokePathHelper.StartPoint.Y = p;
                                 }
 
@@ -350,8 +350,8 @@ namespace LiveChartsCore
                                    nameof(strokePathHelper.StartPoint.Y), nameof(strokePathHelper.StartPoint.X));
                             }
 
-                            strokePathHelper.StartPoint.X = data.X0;
-                            strokePathHelper.StartPoint.Y = data.Y0;
+                            strokePathHelper.StartPoint.X = (float)data.X0;
+                            strokePathHelper.StartPoint.Y = (float)data.Y0;
                             strokePathHelper.Path.AddCommand(strokePathHelper.StartPoint);
                         }
 
@@ -683,12 +683,12 @@ namespace LiveChartsCore
                     Y2 = yScale.ToPixels(next.PrimaryValue + nys),
                     OriginalData = new BezierData(points[i])
                     {
-                        X0 = (float)x0,
-                        Y0 = (float)y0,
-                        X1 = (float)c2X,
-                        Y1 = (float)c2Y,
-                        X2 = (float)next.SecondaryValue,
-                        Y2 = (float)(next.PrimaryValue + nys),
+                        X0 = x0,
+                        Y0 = y0,
+                        X1 = c2X,
+                        Y1 = c2Y,
+                        X2 = next.SecondaryValue,
+                        Y2 = next.PrimaryValue + nys,
                     }
                 };
             }
